Fix #146 PInvoke types should be strctures and not classes
- Have only 1 PInvoke Class
- Optimise enums usage when needed
- Add more native functions and wrappers as per examples in the web.

+++ Source/Krypton Components/ComponentFactory.Krypton.Ribbon/View Draw/ViewDrawRibbonContextTitle.cs
@@ -440,7 +440,7 @@ namespace ComponentFactory.Krypton.Ribbon
                 IntPtr mDC = PI.CreateCompatibleDC(gDC);
 
                 PI.BITMAPINFO bmi = new PI.BITMAPINFO();
-                bmi.biSize = Marshal.SizeOf(bmi);
+                bmi.biSize = (uint) Marshal.SizeOf(bmi);
                 bmi.biWidth = ClientWidth;
                 bmi.biHeight = -ClientHeight;
                 bmi.biCompression = 0;
